requirements.txt
fastapi
uvicorn
openai
python-dotenv
mysql-connector-python

requests==2.31.0
celery
redis
